Add Endpint Ordenes sin calculo comision

+++ AppService.Core/EntitiesMooreve/PcOrdenesSinCalculoComision.cs
@@ -5,6 +5,11 @@ namespace AppService.Core.EntitiesMooreve;
 public class PcOrdenesSinCalculoComision
 {
     public decimal Id { get; set; }
-    public decimal? Orden { get; set; }
-    public string? SearchText { get; set; } 
+
+ 
+    public long? Orden { get; set; } 
+
+  
+    public string? SearchText { get; set; }
+
 }
